+++ util.c
@@ -14,7 +14,7 @@ void ran_seed(long j){
   vran = vran * 2685821657736338717LL;
 }
 
-double ran_ran2(){
+float ran_ran2(){
     vran ^= vran >> 21; vran ^= vran << 35; vran ^= vran >> 4;
     ullong t = vran * 2685821657736338717LL;
     return 5.42101086242752217e-20*t;
@@ -24,7 +24,7 @@ double ran_ran2(){
 //====================================================
 // neighbor list helper functions
 //====================================================
-inline void coords_to_index(double *x, int *size, int *index, double L){   
+inline void coords_to_index(float *x, int *size, int *index, float L){   
     index[0] = (int)(x[0]/L  * size[0]);
     index[1] = (int)(x[1]/L  * size[1]);
 }
@@ -47,6 +47,6 @@ inline int mod_rvec(int a, int b, int p, int *image){
 //==========================================
 // random things
 //==========================================
-inline double mymod(double a, double b){
+inline float mymod(float a, float b){
   return a - b*(int)(a/b) + b*(a<0);
 }
